# docs/swagger.yaml
basePath: /hr-api
definitions:
...
  hr.CreateHrAdminReqModel:
    properties:
      branch_id:
        type: integer
      created_by:
        type: integer
      email:
        type: string
      password:
        type: string
      role:
        type: string
      status:
        type: string
      updated_by:
        type: integer
      user_name:
        type: string
    type: object
...
  hr.CreateServicesReqModel:
    properties:
      category:
        type: string
      updated_by:
        type: integer
      value:
        type: string
    type: object
...
  hr.GetAdminReqModel:
    properties:
      limit:
        type: integer
      page:
        type: integer
      search:
        type: string
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: This is a sample server Petstore server.
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /admin:
    post:
      consumes:
      - application/json
      description: Create a new admin
      parameters:
      - description: Admin details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/hr.CreateHrAdminReqModel'
      produces:
      - application/json
      responses:
        "200":
          description: Admin created successfully
          schema:
            type: string
        "400":
          description: bad request
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Create Admin
      tags:
      - admin
  /admin/all:
    get:
      consumes:
      - application/json
      description: Get all admins
      parameters:
      - description: search query
        in: query
        name: search
        type: string
      - description: page number
        in: query
        name: pageNumber
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hr.GetAdminReqModel'
        "400":
          description: bad request
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Get all Admins
      tags:
      - admin
...
  /service:
    post:
      consumes:
      - application/json
      description: Create a new service
      parameters:
      - description: Service details
        in: body
        name: service
        required: true
        schema:
          $ref: '#/definitions/hr.CreateServicesReqModel'
      produces:
      - application/json
      responses:
        "200":
          description: Service Created
          schema:
            type: string
        "400":
          description: bad request
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Create Service
      tags:
      - services
  /service/{category}:
    get:
      consumes:
      - application/json
      description: Get a service
      parameters:
      - description: service category
        in: path
        name: category
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Service fetched successfully
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Get Service
      tags:
      - services
  /services:
    get:
      consumes:
      - application/json
      description: Get all services
      produces:
      - application/json
      responses:
        "200":
          description: Services fetched successfully
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Get Services
      tags:
      - services
...
swagger: "2.0"